Refactor restore readiness collaborators
